requirements.txt
logzio-python-handler
requests